+++ lib/pages/constant_page.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../utils/loading.dart';
-import '../utils/logger.dart';
+import '../services/app/logger.dart';
+import '../widgets/loading.dart';
 
 /// MaterialApp.router の builder で返すことによって
 /// すべてのページ上に ConstantPage を 挿入することができるので
