+++ lib/services/rest_api/dio/request_interceptor.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 import 'package:dio/dio.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../../utils/logger.dart';
+import '../../app/logger.dart';
 
 final requestInterceptorProvider = Provider<RequestInterceptor>(
   (_) => RequestInterceptor(),
