+++ test/pages/constant_page_test.dart
@@ -3,7 +3,7 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:trip_app_nativeapp/config/app.dart';
 import 'package:trip_app_nativeapp/pages/constant_page.dart';
 import 'package:trip_app_nativeapp/pages/home_page.dart';
-import 'package:trip_app_nativeapp/utils/loading.dart';
+import 'package:trip_app_nativeapp/widgets/loading.dart';
 
 void main() {
   testWidgets(
